Move stubs to utils/tubs
