+++ test/type_binding.test.ts
@@ -2,7 +2,7 @@
 
 import { expect } from 'chai';
 import { TypeBinding, TypeBindingScopeEnum } from "../source/inversify";
-import * as Stubs from './stubs';
+import * as Stubs from './utils/stubs';
 
 describe("TypeBinding", () => {
